Work around for telephony service crashing under FBE
b/29358683 is suspected to be a bug in file based encryption.
UserManger.isUserUnlocked() is reporting true even if the phone is still
locked, and subsequent operations that depends on being unlocked will
crash.

This is a temporary workaround until the issue is resolved. Note that
after this change telephony service is no longer crashing, but the
phone will be stuck in airplane mode when locked so something else is
still wrong.

Change-Id: Ibe5b6f3bdf072b87aa482485d1c6909863dcf1c9