Fix telephony service crashing on CryptKeeper mode
This CL reverts the work around in ag/1154152

CryptKeeper mode is the pre-file based encryption locked state, when the
user has selected "Require password to boot" and the device hasn't been
unlocked yet during a reboot. UserManager.isUserUnlocked() will still
return true in this mode, but storage in /data and all content providers
will not be available(including SharedPreference).

With the possibility of no storage being available during locked state,
the method in ag/884393 which requires storing subIds will not work.
Instead, all subIds are sent through voicemail activation again when the
device is unlocked. This will generate a additional activation request
if file based encryption is not enabled, but duplicated requests are
already an issue (b/28730056) and will be resolved in a future CL.

Change-Id: I75753f67b8598d64c33c83cb6598497664ba4acd
Fixes: 29358683