Disable visual voicemail service on secondary user
When a new user is added, an additional instance of telephony service
will be run, and it will receive broadcasts such as BOOT_COMPLETED.

VVM services is not meant to be run as secondary user, as it is tied
to the phone account. The Phone object it rely on is also not available
to secondary users.

In this CL, all VVM broadcast receivers will only run on system user.

Bug: 29574016
Change-Id: I1408de32ca146e210f6b27c3a7c633deb33813dc