Set message count to 0 while clearing MWI
The Message Waiting Indicator is the traditional way of informing the
phone there's some voicemail waiting. The carrier will sent an special
SMS containing the current unread message count, which is intercepted
by telephony. Telephony will show a notification if the unread count is
greater then zero.

With a active visual voicemail, this feature should be suspended to
avoid duplicated notification.

Previously only the notification itself is hidden, but the unread count
is cached in the system and notification will show again when the phone
reboot, before visual voicemail is activated. Most carrier also stop
sending MWI SMS if visual voicemail is activated, so this value will be
stuck in the system forever.

In this CL instead of just hiding the notification, the message count
itself will be wiped to ensure the notification will never be shown
again unless VVM is disabled and another MWI SMS is received.

Change-Id: I1ebe7fc3505833c45a15a386ab93b54afbfe1be1
Fixes: 29122194