Handle non-ready provisioning status on non-VVM3
Currently only VVM3 has implemented provisioning steps.
Before implementing provisioning we used to ignore the provisioning
status and attempt to access the server. By having all protocol
going through the provisioning process non-VVM3 protocols will do
nothing and not complete the activation.

In this CL if provisioning is not supported for the protocol, the old
behavior will be used. Meaning the provisioning status will once again
be ignored.

Change-Id: Iebf9c159bc4a1c4fe91d5734f5f53c6d90c0fa8a
Fixes: 30063031