Key the conference participants using the Endpoint
We currently keep track of the conference participants using the address
from the Conference Event Package. This can potentially be the same if
two or more of the participants include connections to the same
device. We will now use the endpoint as the unique identifier to be sure
that there are no possible duplicates.

Change-Id: Ice7d7b09cdb5fdcc844e1ab89cf255dffeca2b9e
Fixes: 30089057