Change how VoWifi status and wifi status are calculated.
Check to see whether VoWifi is enabled by querying the user setting
instead of the ImsManager.
Check to see whether the current call is a wifi call by reading the
property inside TelephonyConnection instead of querying the ImsCall.
Update whether a call can be added when a call switches from non-wifi to
wifi or vice versa.

Bug: 29047863
Change-Id: Ifc1d48b9a1ba3b974bcd1885c8566defcf7c5c42