Facilitate Hanging Up Pending Emergency Call
Corrects the hangup logic for calls that have not successfully returned
a telephony Connection. Before, we only used to hang up calls that were
backed by a telephony Connection. This made calls hang that were waiting
for the telephony Connection to be established (such as Emergency
Calls or calls with bad service).

Change-Id: Id8a0a5c4e04736e9697323a7fc5bf293fd09e343
Fixes: 29433202