Show voicemail notification under FBE
Under FBE we have no access to stored credentials so interacting with
the server is impossible. But the new message SYNC SMS does provide us
enough information to show the voicemail notification.
LegacyModeSmsHandler is designed to show notifications purely based on
SMS, and is suited for this task.

After the device is unlocked, VVM will be activated and dismiss the
notification. Another notification will be shown after dialer synced
the new messages.

Change-Id: Id9a1bfb6247389a01c809a6f61fe3f7dde8fdfa1
Fixes:30102262