Fix issue where the carrier presence cap. may not be set on phoneaccounts.
CAPABILITY_VIDEO_CALLING_RELIES_ON_PRESENCE is only set if the phone
account also supports video.  If when the phone account is originally
created, there is no video capability, the presence bit will not be set.
The old code which handles changes to video capability after initial
creation of the phone account assumed that presence was set on the
phone account so that it knew to set it again.  However if it was never
set in the first place we'd fail to set it again.

Added method to TelecomAccountRegistry to query for cached carrier config
value, and then ensure that the presence capability is being set
accordingly.

Bug: 30225766
Change-Id: I3bea2af86900e265db24e248eca6ef71e5e5b79d