Wipe data channel state when signal returns
When the signal is lost and cellular data is required for the carrier,
DATA_CHANNEL_STATE_NO_CONNECTION_CELLULAR_REQUIRED will be set.
When the signal is returned, the error was not cleared immediately.
Instead, it was cleared when a sync task is started. Although a sync
task usually follows the signal return, it could be blocked by the
sync interval requirement, and the error state will remain forever.

After this CL, the data channel state is reset after the signal
returns. If there are any errors it would be posted again by the sync
task.

Change-Id: I810eb8b89e19cb0fe5091b4d63e393cc93a8e9d0
Fixes:30263279