Fix issue where multiple anonymous participants don't show up correctly.
On Vzw when there are multiple anonymous participants, the SIP uris
returned in the CEP are different from what was expected ; eg:
sip:anonymous1@anonymous.invalid;legid=1
sip:anonymous2@anonymous.invalid;legid=2

The CEP parsing code did not handle the ; and extra parameter.
Per RFC3261, when we parse this string:
anonymous1@anonymous.invalid;legid=1
We first separate out and ; parameters (which are guaranteed to be at
the end of the address.

Bug: 30434469
Change-Id: I2f7224f434a0b01f54d6815276a8bb365bd53ac1