Reset voicemail status to disabled for not supported phone accounts
In some cases a phone account might have already written to the status
table, but the feature is later removed through changes in
CarrierConfig. Previously the VVM client will just stop handling the
account and the status will be stuck in the table. Sometimes it would
leave a error that which will be shown to the user forever.

After this CL, if the phone account has no VVM support, the status
table entry will be reset to the disabled state, which will make the
UI ignore the account.

Bug: 30314643
Change-Id: I0b67decd62d021de90d7d0ff0f5209e7f3e624ee