Tweak visual voicemail
b/30513337 NPE when connection lost
ImapConnection will try to log out when the connection closes, even
when the connection is interrupted and logging out cannot be done.
Doing so will cause a NPE.

In this CL log out will only commence if the transport is still valid.

b/30514504 Don't show "Activating" if VVM has been activated before
The activating message is shown upon every boot but most of the time
the account should be already activated, activation is just for book
keeping.

After this CL the activating message will only be shown if the account
is never activated before on the device. Error message will still be
posted if activation fails.

b/30483028 getCallCapablePhoneAccounts() returns null
This is working as intended, comments updated

b/29363739 Unable to fetch voicemail from Marshmallow
PhoneAccountHandle is stored in the database to determine which account
to fetch voicemail from. Between M and N the PhoneAccountHandle value
has changed, causing the entry in M voicemails to mismatch.

In this CL if the phone account of the voicemail cannot be found, the
client will try to search it as the M format

+ Fix NPE while updating quota when the preceding operations on the
  ImapHelper has already failed.

Change-Id: Ia17e345c772892e17a005ae8f0c6c08255a0ef1f
Fixes: 29363739
Fixes: 30513337
Fixes: 30514504