Replace transcription for unsupported VVM message type
The VVM client only supports audio messages. Most other message types
are filtered the moment the SYNC SMS is received, but some message type
such as VVM3 DSN is indistinguishable from audio message and will pass
through.

DSN (Delivery Status Notification) is the notification when sending a
email to some address has failed. VVM3 DSN triggers a SYNC SMS that is
completely identical to normal audio messages, and can only be
identified after accessing the IMAP server.

Although DSN is not supported in the VVM client, some carrier has this
feature in their voicemail telephone interface, and will send the DSN
to the client, this must be handled or a unplayable message will be
shown with no indication of errors.

In this CL, if the client is unable to find the audio data for a
voicemail, it will be deemed a an unsupported message, and its'
transcription field will be replaced by an error message. This covers
not only DSN but also any other unplayable message.

While this solution is far from ideal, it is not expected to have a lot
of users that use the DSN feature or other unknown message types.

Change-Id: I482479b746776b563ebf5694a58a8c2cf534b0f2
Fixes: 30588635