Fix MWI missing on blocked VVM3 SIM
Previously MWI will be suppressed if the notification channel state
(phone signal) is active. This will only work if VVM is fully active
or fully disabled. In VVM3 there is a new state that the VVM client is
enabled but the carrier has blocked the service. The notification
channel state will be active, but VVM SMS will not be sent and MWI will
continue as usual and mistakenly be suppressed by VVM client.

After this CL, instead of checking the notification state to suppress
MWI, the configuration state which determines if the service is actually
active will be used.

Change-Id: Id55a88ff3c188113f1f40f56503d2b61dd5c0c40
Fixes: 30883795