Add exception handling and retry when configuring VVM.
In VoicemailStatus#apply there is a possibility for the call to
ContentResolver#insert(.., ...) to throw an IllegalArgumentException if
the content resolver can't be started.  If this happened there was a
possibility of the entire Phone process dying.

Added exception handling to the apply method, and in the
OmtpVvmCarrierConfigHelper added code to retry configuration if it fails.

Bug: 29837276
Change-Id: I419e866e0b44b69932f4d0b9d2b46067ce2efcfe