Fix VVM Activation
In ag/1337726 ActivationTask.mData was misused to carry the flag to
register the content providers. mData's purpose is solely to store the
data from the status SMS, and if it is null a new status SMS will be
requested. Putting the content provider flag in it caused the 0status
SMS never to be requested and broke the activation process.

In this CL, docs and field names are updated to clarify this, and the
content provider registration will be done on every activation since it
should usually be a relatively cheap process.

Fixes: 31350221
Change-Id: I36cc08cac60df77c2b77dc7fdb5a96eeaf8e8052