Cleanup Connections in conference controllers on CDMA<-->GSM change.
When a call switches technology between CDMA and GSM, the
TelephonyConnectionService will end up re-adding it to the appropriate
conference controller due to the original connection change.  This is
problematic, as it means that the connection will still be tracked by the
conference controller of its previous technology.  Worse yet, if a call
makes a GSM-->CDMA-->GSM transition, it will be added again to the
GSM call tracker.

There is still an underlying issue in that the TelephonyConnection subclass
type doesn't change when the radio technology changes -- that is a larger
issue to solve and should be addressed at some point (ie not now).

To fix this, added code to ensure that connections are removed from other
conference controllers when added, and also to prevent adding duplicates.

Bug: 31288401
Change-Id: I5e95d4f72be6af473eeec1b92f47ad77a783c9b2