Optimize visual voicemail activation
Before this CL VVM voicemail will wait for a 60 seconds timeout for the
carrier to respond the status request, Regardless of the situation.

After this CL if the cellular signal is lost before the task is ran,
the activation task will abort immediately. Failing to send the SMS
will also fail the task immediately and queue for a retry.

This change should significantly reduce the time the wake lock is held
waiting for the status response when the signal is poor.

Change-Id: I8284d00b52875ba067afdda21dc3dbbb02f103bd
Fixes: 30126770
Bug: 31049275