USSD: fix USSD request pop up “unable to process request” issue
This patch is to fix a bug that MMI Complete case still displays
"unable to process request". In the original logic, the Complete
case always fallback to the Failed case to display the message
got through mmiCode.getMessage(). Hence add state check to make
sure the notification is displayed for Failed case

Change-Id: I09e88b11d4e73cd9b572081f00fa234c3893b3b5
CRs-Fixed: 1066212