Tweak change voicemail PIN UI
b/31442531 Able to submit a mismatch pin during Voicemail PIN change
PIN mismatch was not checked when pressing the next button on the IME.
Fixed.

b/30818616 VVM Change PIN UI is cropped when Font size is set to large
Removed the bottom padding of the container. The container already
tries to take as much as vertical space as it can, and this will free
up some space for the EditText when the font is large.

b/31224822 Cannot activate VVM after entering wrong PIN a few times
The "change PIN" setting will now only appear after there are
successful activation records.

b/31065311 "Change PIN" for Voicemail is available when VVM is disabled
The "enabled" status was not rechecked when the activity is resumed.
Fixed.

- Deleted voicemail_dialog_change_pin.xml
  This is the layout for the old change PIN UI, no longer used.
- Removed VisualVoicemailSettingsUtil.isEnabled(Phone phone)
  PhoneAccountHandle cached in VoicemailSettingsActivity, no longer
  needed this overload.

Change-Id: I3f649cbd4131d5f0262c7d5b8cc2297e4dfb579d
Fixes: 31442531
Fixes: 30818616
Fixes: 31224822
Fixes: 31065311