Implement long sleep for VVM task scheduler
If the next task is more than 60 seconds away, instead of using
Handler.postDelayed() TaskScheduler will set the AlarmManager and
release the wakelock. This will allow longer sleeps without draining
the battery.

Currently only the SyncTask is affected by this as it has a 60 seconds
retry. ActivationTask and SyncTask are expected to add a exponential
retry in the future that will use the long sleep more.

Bug: 30252737
Change-Id: I8b1f3fe28f49a40abf45bb0a56e3fad2ed243ed8