Restore VVM config for Orange France
This CL reverts the Orange France related changes in ag/1375743

Orange France has requested the VVM config to be removed due to
b/28868979 not being fixed in previous release (VVM should be disabled
when it shut itself down due to the carrier VVM app being installed).
As the bug is fixed the config should be restored.

The other removed config marked as "Orange UK" in ag/1375743 was for
EE and they operate separately. It was added in ag/690384 for testing
and never fully supported. It should remain removed.

Note that the values in CarrierConfig are not restored. In VVM
CarrierConfig is only used for carrier runtime overrides, the default
values are put in vvm_config.xml.

Change-Id: I5d2cabc9ce60c6f62204c960b9fcfd2bf3a492aa
Fixes: 31664337