Mark ACTION_REQUEST_SENT_INTENT as protected.
The ActivityManager was complaining about sending a non-protected
broadcast from the system process.

StatusSmsFetcher is a dynamic receiver so it cannot receive a
component specific intent. b/31749921 also caused the PendingIntent to
be fired with no receiver and triggered
ActivityManager.checkBroadcastFromSystem().

In this CL the broadcast is marked as protected and broadcasted to the
telephony package so the above issues can be avoided.

Fixes: 31700392
Change-Id: I420b9e782892edc7b873c2adea1e96be063cf80e