Allow initial voicemail status setup without signal
During device shutdown the VVM status might be wiped as the SIM will be
perceived as removed, setting the configuration state as "not
configured". During this state dialer will not show the VVM tab. The tab
will be restored when activation is ran because the configuration state
is set back to OK or ACTIVATING.

During previous optimization, the activation will not be queued at all
if the phone does not have signal. In this case after a shut down the
user will not see the voicemail tab again until signal is restore and
activation is ran.

In this CL, VVM activation will be queued even if there are no signal
the activation task will do the initial status table setup to bring
back the VVM tab,then check for signal and abort. Notification channel
state is wiped during the setup, but the "no signal" will be restored
if the task is aborted due to no signal.

Change-Id: Ib7c376b0a208fc14452ece0f5fad84b136f97d45
Fixes: 31903828