Remove source when disabling VVM
The source is a PhoneAccountHandle is used to track which accounts
currently has VVM activated. It should be removed when VVM is disabled
or reactivation will be prevented.

In b/32371710 there's a huge torrent of
VvmPhoneStateListener.onServiceStateChanged causing it to determine
the carrier is not supported. VVM is shut off, the tab was hidden,
but the source is not removed.

Later when the signal returned, because the source is still there,
reactivation did not happen and the tab is not restored.

Change-Id: I9dc6f7df84c68eb3257a170492e6a219e5f1d17d
Fixes: 32371710
Test: make