Preserve PIN_NOT_SET state when reactivating on VVM3
In MR1 if the account has been activated on the device before, during
reactivation the status will be wiped clean as it is just book keeping.
The PIN_NOT_SET flag is also removed, so the user will not see the
warning before the reactvation is done, opening a window to decline ToS
and disable VVM without being warned the PIN has change after reboot.

This CL checks the PIN status during reactivation.

BUG: 33220569
Change-Id: I79fe08a31720315fb52ccbdef714ad8155b71425
FIXES: 33220569
TEST: manual test: set prefernce "default_old_pin" and reboot