Initiate call through user initiated sub
When sim is removed and user tries with the same
sub to initiate call, TeleService fetch the phone
with the same sub id. But the sub is already removed.
So, the subId will be invalid. PhoneFactory will
return default phone when sub id will be invalid.
So, call will be initiated through sub0.

Call should be always initiate through the user
initiated sub. So, TeleService should fetch the
subId from requested AccountHandle.

Change-Id: I86772edf2a54e133aa8ba0fc9a47f23141b2390b
CRs-Fixed: 2001056