Fix the secondary call info is shown when hanging up conference call
Telephony connection does not take state DISCONNECTING, so even the
original connection is in disconnecting state, the child call of
the conference call in upper layer is still active, because the parent
call of the child call becomes null, UI always gets an active call to
show as secondary call.

To fix it, remove the capability HOLD when disconnecting so that UI skips
showing the secondary info with an active child call.

Change-Id: Ibc4db292523ca9205eb12478bd1898fa65f1e66b
CRs-Fixed: 2012580